==> Flu_analysis_models/02_Logistic_KNN_SVC_Flu.ipynb <==
# imports
from sklearn.model_selection import cross_val_score, GridSearchCV
from sklearn.linear_model import LogisticRegression
from sklearn.impute import SimpleImputer
from sklearn.preprocessing import OneHotEncoder, StandardScaler
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline
from sklearn.metrics import f1_score
from sklearn.neighbors import KNeighborsClassifier
from sklearn.svm import SVC

import pandas as pd
import numpy as np
import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)
%matplotlib inline

X_train = pd.read_csv('../data/X_train_flu.csv', index_col= 'respondent_id')
X_test = pd.read_csv('../data/X_test_flu.csv', index_col= 'respondent_id')
y_train = pd.read_csv('../data/y_train_flu.csv', index_col= 'respondent_id')
y_test = pd.read_csv('../data/y_test_flu.csv', index_col= 'respondent_id')

y_train = y_train['seasonal_vaccine']
y_test = y_test['seasonal_vaccine']

X_train.head()

X_train.isna().sum()

y_train.isna().sum()

X_train_cat = X_train.select_dtypes('object')
X_train_num = X_train.select_dtypes(['float64', 'int64'])

cat_pipe = Pipeline([
    ('impute', SimpleImputer(strategy='most_frequent')),
    ('ohe', OneHotEncoder(sparse=False, handle_unknown='ignore'))
])
num_pipe = Pipeline(steps=[
    ('impute', SimpleImputer(strategy='most_frequent'))
])

transformer = ColumnTransformer([('categorical', cat_pipe, X_train_cat.columns),
                                 ('numerical', num_pipe, X_train_num.columns)])

logreg_pipe = Pipeline([
    ('trans', transformer),
    ('logreg', LogisticRegression(n_jobs=-1, random_state=42))
])
logreg_pipe.fit(X_train, y_train)

logreg_cv = cross_val_score(logreg_pipe, X_train, y_train, n_jobs=-1, verbose=3)

logreg_cv.mean()

# Logistic Regression with Grid Search

logreg_pipe_grid = {'logreg__C': [1e-2, 1, 1e2]}
gs_logreg_pipe = GridSearchCV(estimator=logreg_pipe, param_grid=logreg_pipe_grid,
                              cv=5, verbose=3, n_jobs=-1)
gs_logreg_pipe.fit(X_train, y_train)

gs_logreg_pipe.best_params_

log_cv_f1 = cross_val_score(gs_logreg_pipe.best_estimator_, X_train, y_train, n_jobs=-1, verbose=3) 

log_cv_f1.mean()

With a Logistic Regression model, I am able to get an accuracy score of $0.776$. All future models will have to have a score higher than this one in order to be viable models.

# KNN

knn_pipe = Pipeline([
    ('trans', transformer),
    ('knn', KNeighborsClassifier(n_jobs=-1))
])
knn_pipe.fit(X_train, y_train)

knn_pipe_cv = cross_val_score(knn_pipe, X_train, y_train, n_jobs=-1, verbose=3)

knn_pipe_cv.mean()

## KNN with a Grid Search

knn_pipe_grid = {'knn__n_neighbors': [5, 7, 9], 'knn__p': [1, 2, 3]}
gs_knn_pipe = GridSearchCV(
    estimator=knn_pipe, param_grid=knn_pipe_grid, cv=5, n_jobs=-1, verbose=3)
gs_knn_pipe.fit(X_train, y_train)

gs_knn_pipe.best_params_

knn_pipe_cv = cross_val_score(gs_knn_pipe.best_estimator_, X_train, y_train, n_jobs=-1, verbose=3)

knn_pipe_cv.mean()

KNN was not better than a logistic regression when getting the best accuracy score of the test data. With an accuracy score of $0.737$, this model is lower than the logistic regression model that had an accuracy of $0.776$.

# Support Vector Classifier

svm_pipe = Pipeline([
    ('trans', transformer), 
    ('svm', SVC())
])
svm_pipe.fit(X_train, y_train)

svc_cv_score = cross_val_score(svm_pipe, X_train, y_train, n_jobs = -1, verbose = 3)

svc_cv_score.mean()

# SVC with a Grid Search

svm_pipe_grid = {'svm__C': [.001, 1, 100, 1000],
                 'svm__gamma': [0.001, 1, 100]}

gs_svm_pipe = GridSearchCV(estimator=svm_pipe, param_grid=svm_pipe_grid, verbose=1, cv=5, n_jobs=-1)
gs_svm_pipe.fit(X_train, y_train)

gs_svm_pipe.best_params_

svc_cv_score = cross_val_score(gs_svm_pipe.best_estimator_, X_train, y_train, n_jobs = -1, verbose = 3)

svc_cv_score.mean()

The final model that will be made in this notebook will be an SVC. After running a grid search on this model I am able to get a score of $0.780$ which is higher than both the logistic regression and the KNN Classifier. In future notebooks, I will create a Random Forest Classifier, Graidant Boost Classifier, XGBoost Classifier, and a Neural Notebook.
